+++ libs/vellum-compiler/vellum/vellum.h
@@ -3,6 +3,8 @@
 #include <filesystem>
 #include <vector>
 
+#include "common/types.h"
+
 namespace fs = std::filesystem;
 
 namespace vellum {
@@ -10,6 +12,7 @@ class Vellum {
  public:
   bool run(const fs::path& inputFile,
            const std::vector<fs::path>& importPaths,
-           bool emitDebugInfo = true);
+           bool emitDebugInfo = true,
+           common::Opt<fs::path> outputDirectory = std::nullopt);
 };
 }  // namespace vellum